Change "Working hours" by "Turnaround time"
Working hour was defined as the time a project was completed minus the
time it was assigned. This does not represent the real working hour that
the reviewer dedicated to the review.

As turnaround time is defined as the total time taken between the submission
of a program/process/thread/task for execution and the return of the complete
output to the customer/user, it seems more appropriate.